requirements.txt
dask==2025.4.1
pyiceberg==0.9.0
pyarrow==20.0.0
papermill==2.6.0
s3fs==2025.3.0
fsspec==2025.3.0
clickhouse_connect==0.8.15
dotenv
numpy
pandas==2.3.2
pyyaml==6.0.2
jupyter==1.1.1